requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
httpx==0.25.2
pytest-cov==4.1.0
